So many small bugs:(
The Chase system still has many problems for both ground and flying enemies.
Flying - A* algorithm works perfectly but enemies are big, and they 'graze' the sides of obstacles causing wierd movements when chasing the player

Ground - A* algorithm does not work well with the fact that ground enemies cannot move through air. Either another algorithm will be made, or some conditions in the A* algorithm must be added - the question is which for best performance